+++ app/controllers/user/decks/index.js
@@ -1,5 +1,6 @@
 import Controller from '@ember/controller';
 import { computed } from '@ember/object';
+import RSVP from 'rsvp';
 
 export default Controller.extend({
 	queryParams: ['cardset', 'class'],
@@ -70,7 +71,9 @@ export default Controller.extend({
 				user: user,
 				deck: deck
 			});
+			this.set('lock', true);
 			wanteddeck.save().then(wanteddeck => {
+				let wantedcardPromises = [];
 				deck.deckcards.forEach(deckcard => {
 					let userCollection = deckcard.card.get('collections').filter(collection => {
 						return collection.user.get('id') === user.id;
@@ -84,7 +87,7 @@ export default Controller.extend({
 								wanteddeck: wanteddeck,
 								number: missingNumber
 							});
-							wantedcard.save();
+							wantedcardPromises.push(wantedcard.save());
 						}
 					} else {
 						const wantedcard = this.get('store').createRecord('wantedcard', {
@@ -93,27 +96,32 @@ export default Controller.extend({
 							wanteddeck: wanteddeck,
 							number: deckcard.number
 						});
-						wantedcard.save();
+						wantedcardPromises.push(wantedcard.save());
 					}
+				});
+				RSVP.allSettled(wantedcardPromises).then(() => {
+					this.set('lock', false);
 				});
 			});
 		},
 
 		removeWanteddeck(wanteddeck) {
-			wanteddeck.get('wantedcards').then(wantedcards => {
-				wantedcards.forEach(wantedcard => {
-					wantedcard.get('card').then(card => {
-						card.wantedcards.removeObject(wantedcard);
+			if (!this.get('lock')) {
+				wanteddeck.get('wantedcards').then(wantedcards => {
+					wantedcards.forEach(wantedcard => {
+						wantedcard.get('card').then(card => {
+							card.wantedcards.removeObject(wantedcard);
+						});
+						wantedcard.deleteRecord();
+						wantedcard.save();
 					});
-					wantedcard.deleteRecord();
-					wantedcard.save();
 				});
-			});
-			wanteddeck.get('deck').then(deck => {
-				deck.wanteddecks.removeObject(wanteddeck);
-			});
-			wanteddeck.deleteRecord();
-			wanteddeck.save();
+				wanteddeck.get('deck').then(deck => {
+					deck.wanteddecks.removeObject(wanteddeck);
+				});
+				wanteddeck.deleteRecord();
+				wanteddeck.save();
+			}
 		}
 	}
 });
